refactor: move aws auth logic into a common module
