+++ bucket-blocker/main.go
@@ -7,6 +7,7 @@ import (
 	"github.com/aws/aws-sdk-go-v2/service/cloudformation"
 	"github.com/aws/aws-sdk-go-v2/service/s3"
 	"github.com/aws/aws-sdk-go-v2/service/securityhub"
+	awsauth "github.com/guardian/fsbp-tools/aws-auth"
 	"github.com/guardian/fsbp-tools/bucket-blocker/utils"
 )
 
@@ -15,7 +16,7 @@ func main() {
 	ctx := context.Background()
 	args := utils.ParseArgs()
 
-	cfg, err := utils.LoadDefaultConfig(ctx, args.Profile, args.Region)
+	cfg, err := awsauth.LoadDefaultConfig(ctx, args.Profile, args.Region)
 	if err != nil {
 		log.Fatalf("unable to load SDK config, %v", err)
 	}
