+++ bucket-blocker/utils/awsutils.go
@@ -3,45 +3,18 @@ package utils
 import (
 	"bufio"
 	"context"
-	"errors"
 	"fmt"
 	"os"
 	"strings"
 
 	"github.com/aws/aws-sdk-go-v2/aws"
-	"github.com/aws/aws-sdk-go-v2/config"
 	"github.com/aws/aws-sdk-go-v2/service/cloudformation"
 	cfnTypes "github.com/aws/aws-sdk-go-v2/service/cloudformation/types"
 	"github.com/aws/aws-sdk-go-v2/service/s3"
 	s3Types "github.com/aws/aws-sdk-go-v2/service/s3/types"
 	"github.com/aws/aws-sdk-go-v2/service/securityhub"
 	shTypes "github.com/aws/aws-sdk-go-v2/service/securityhub/types"
-	"github.com/aws/aws-sdk-go-v2/service/sts"
 )
-
-func validateCredentials(ctx context.Context, stsClient *sts.Client, profile string) (*sts.GetCallerIdentityOutput, error) {
-	resp, err := stsClient.GetCallerIdentity(ctx, &sts.GetCallerIdentityInput{})
-	if err != nil {
-		return resp, errors.New("Could not find valid credentials for profile: " + profile)
-	}
-	return resp, nil
-}
-
-func LoadDefaultConfig(ctx context.Context, profile string, region string) (aws.Config, error) {
-	cfg, err := config.LoadDefaultConfig(ctx, config.WithSharedConfigProfile(profile), config.WithDefaultRegion(region))
-	if err != nil {
-		fmt.Println("Error loading configuration")
-		return cfg, err
-	}
-
-	stsClient := sts.NewFromConfig(cfg)
-	_, err = validateCredentials(ctx, stsClient, profile)
-	if err != nil {
-		return cfg, err
-	}
-
-	return cfg, nil
-}
 
 func findFailingBuckets(ctx context.Context, securityHubClient *securityhub.Client, bucketCount int32) ([]string, error) {
 	controlId := "S3.8"
