+++ go.work
@@ -3,3 +3,5 @@ go 1.22.1
 use ./bucket-blocker
 
 use ./ingress-inquisition
+
+use ./aws-auth
